some fixes

+++ src/controllers/movementations-controller.js
@@ -10,25 +10,16 @@ const move_type = {
 const movementations = async (req, res) => {
   try {
     const userId = req.params.id
-    const employer = await prisma.user.findUnique({
+    const user = await prisma.user.findUnique({
       where: {
         id: userId,
       },
     })
 
-    const employee = await prisma.employee.findUnique({
-      where: {
-        id: userId,
-      },
-    })
-
-    const user = employer ? employer : employee
-    const parentItemId = user.user_id ? user.user_id : user.id
-    
     if (user) {
       const movementations = await prisma.movementation.findMany({
         where: {
-          user_id: parentItemId,
+          user_id: user.id,
         },
         include: {
           user: true,
@@ -38,7 +29,7 @@ const movementations = async (req, res) => {
       })
       res.status(200).json(movementations)
     } else {
-      res.status(404).json({ message: 'Loja não encontrada' })
+      res.status(404).json({ message: 'Estabelecimento não encontrado' })
     }
   } catch (error) {
     console.error('Error:', error)
@@ -52,7 +43,7 @@ const move = async (req, res) => {
     const itemId = body.itemId
     const userId = body.userId
 
-    if (!itemId && !userId) {
+    if (!itemId || !userId) {
       return res.status(401).json({ message: 'ID do usuário ou do item inválido' })
     } else {
       const employer = await prisma.user.findUnique({
@@ -76,7 +67,7 @@ const move = async (req, res) => {
       let newQuantity = 0
       if (body.move_type != move_type.entrada) {
         if (body.quantity > item.quantity) {
-          return res.status(401).json({ message: 'A Quantidade do item é inferior ao solicitado' })
+          return res.status(401).json({ message: 'A quantidade do item em estoque é inferior ao solicitado' })
         } else {
           newQuantity = item.quantity - body.quantity
         }
@@ -104,7 +95,7 @@ const move = async (req, res) => {
               create: {
                 move_type: body.move_type,
                 user_id: employer ? employer.id : employee.id,
-                quantity: newQuantity,
+                quantity: body.quantity,
                 employee_id: employee?.id,
                 updated_at: new Date().toISOString(),
               }
@@ -133,13 +124,32 @@ const deleteMovementation = async (req, res) => {
     if (!movementation) {
       res.status(404).json({ message: 'Movimentação não encontrada' })
     } else {
-      await prisma.movementation.delete({
+      const itemReference = await prisma.item.findUnique({
         where: {
-          id: movementationId,
+          id: movementation.item_id,
         },
       })
 
-      res.status(200).json({ message: 'Movimentação deletada com sucesso' })
+      if (itemReference) {
+        const newQuantity = movementation.move_type != move_type.entrada ? itemReference.quantity + movementation.quantity : itemReference.quantity - movementation.quantity
+        
+        await prisma.item.update({
+          where: {
+            id: movementation.item_id,
+          },
+          data: {
+            quantity: newQuantity,
+          },
+        })
+        
+        await prisma.movementation.delete({
+          where: {
+            id: movementationId,
+          },
+        })
+
+        res.status(200).json({ message: 'Movimentação deletada com sucesso' })
+      }
     }
   } catch (error) {
     console.error('Error:', error)
